new commit

+++ src/main/java/com/dawood/peeng/monitor/models/MonitorCheck.java
@@ -41,7 +41,7 @@ public class MonitorCheck extends MetaData {
 
   private Integer statusCode;
 
-  private Integer responseTimeMs;
+  private Long responseTimeMs;
 
   @Column(columnDefinition = "TEXT")
   private String errorMessage;
